Use consistent order on molecule index page [4]
https://www.pivotaltracker.com/story/show/163703441 

https://github.com/elastic/elasticsearch-rails/pull/835/files

when viewing molecules on the Explore Data - Molecules 2nd row tab, molecules are show in descending order. However, if we add a search term to query against a batch field/udf, the molecules are no longer sorted in descending order. the result of this issue, is that when users toggle between different # of mols per page, they will see different blocks of molecules in a random order.